backend/requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-SocketIO>=5.3.0
python-socketio>=5.8.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
pymongo>=4.5.0
face-recognition>=1.3.0
dlib>=19.24.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
python-dateutil>=2.8.0 